lab-04/websocket_aes_client_ui/requirements.txt
PyQt5
websocket-client
pycryptodome